feat: Double 타입 BigDecimal 타입으로 변경
- 부동 소수점으로 인한 정확도 문제로 타입 변경

+++ src/main/java/com/mock/investment/stock/domain/account/dto/AccountBalanceDto.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.PositiveOrZero;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * DTO for {@link com.mock.investment.stock.domain.account.domain.Account}
  */
@@ -16,13 +18,13 @@ public class AccountBalanceDto {
 	String accountNumber;
 
 	@PositiveOrZero
-	Double krwBalance;
+	BigDecimal krwBalance;
 
 	@PositiveOrZero
-	Double usdBalance;
+	BigDecimal usdBalance;
 
 	@PositiveOrZero
-	Double bitcoinBalance;
+	BigDecimal bitcoinBalance;
 
 	public static AccountBalanceDto fromEntity(Account account) {
 		return AccountBalanceDto.builder()
