+++ src/main/java/com/mock/investment/stock/domain/order/dto/BuyOrderRequest.java
@@ -7,15 +7,17 @@ import jakarta.validation.constraints.Positive;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class BuyOrderRequest extends OrderRequest{
     @Positive
     @Schema(description = "주문 수량", example = "10")
-    private Double quantity;
+    private BigDecimal quantity;
 
     @Schema(description = "주문 가격", example = "10000")
-    private Double price;
+    private BigDecimal price;
 
     @Schema(description = "주문 타입", example = "LIMIT")
     private OrderType orderType;
